+++ src/preview/languages/stylesheet.ts
@@ -1,7 +1,6 @@
-import axios from "axios";
 import * as path from "path";
-import { store } from "src/store";
-import { TextDocument, Uri, workspace } from "vscode";
+import { TextDocument } from "vscode";
+import { store } from "../../store";
 import { byteArrayToString } from "../../utils";
 
 export const STYLESHEET_BASE_NAME = "style";
@@ -35,22 +34,13 @@ export async function getStylesheetContent(
       case StylesheetLanguage.scss:
       case StylesheetLanguage.sass: {
         const sass = require("sass");
-        return byteArrayToString(
-          sass.renderSync({
-            data: content,
-            indentedSyntax: extension === StylesheetLanguage.sass,
-            importer: async (url: string, previous: string, done: Function) => {
-              if (url.startsWith("http")) {
-                const { data } = await axios(url);
-                done({ contents: data });
-              } else {
-                const uri = Uri.joinPath(store.activeSwing!.currentUri, url);
-                const contents = await workspace.fs.readFile(uri);
-                done({ contents: byteArrayToString(contents) });
-              }
-            },
-          }).css
-        );
+        const { css } = sass.renderSync({
+          data: content,
+          indentedSyntax: extension === StylesheetLanguage.sass,
+          includePaths: [store.activeSwing!.currentUri.path]
+        });
+
+        return byteArrayToString(css);
       }
       case StylesheetLanguage.less: {
         const less = require("less").default;